Fix dependency problems with example/test files.
There was an issue with the dependencies with the test source code. For
some reason the test source files were not being added to the test
program's build even though they were listed in add_executable. I think
there might be a bug in CMake where if you add an external dependency to
a source file, it sometimes removes it from the sources included in the
program build.