Vec classes no longer implicitly convert scalars in multiply
Previously, the vtkm::Vec class supported scalar multiply for all basic
types where the type would be implicitly converted to the Vec component
type. This was removed because these implicit conversions were causing
lots of trouble with conversion warning flags.

One of the examples relied on this implicit conversion. It has been
fixed.