Fixes for subtle changes to FunctionInterface
A recent change was made to FunctionInterface and related functions to
prevent the copying of data. Part of this was to change the templated
parameters of functions like make_FunctionInterface to be constant
references rather than values.

The declaration of the function argument as the const ref of a template
parameter caused the template parameter type to resolve to weird things
for C strings. For some reason when you use a string literal, the const
gets pulled off, but it does not if you use a typed variable or specify
the type.

This change fixes these instances. The odd behavior still remains, but
putting C strings in a FunctionInterface is not really a viable use case
for VTK-m anyway.