Add example of using provided worklet
Originally I was thinking I was going to add documentation for all of
the provided worklet. But now that they are wrapped around filters, that
seems excessive. It is probably just fine to document all the filters,
point folks to the worklets, and let them find it on their own.